requirements.txt
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
arrow==1.3.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==21.2.0
bcrypt==3.2.0
binaryornot==0.4.4
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cookiecutter==2.6.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2.4
django-dotenv==1.4.2
django-jazzmin==3.0.1
django-querycount==0.8.3
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
dotenv==0.9.9
drf-yasg==1.21.10
duplicity==0.8.21
exceptiongroup==1.3.0
fastapi==0.116.0
fasteners==0.14.1
flet==0.28.3
flet-cli==0.28.3
flet-desktop-light==0.28.3
flet-web==0.28.3
future==0.18.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
Jinja2==3.1.6
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.3.1
olefile==0.46
packaging==25.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
psycopg==3.2.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypng==0.20220715.0
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.4.2
redis==6.4.0
repath==0.9.0
reportlab==3.6.8
requests==2.25.1
rich==14.0.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
systemd-python==234
text-unidecode==1.3
texttable==1.6.4
toml==0.10.2
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.35.0
uvloop==0.21.0
wadllib==1.3.6
watchdog==4.0.2
watchfiles==1.1.0
websocket-client==1.2.3
websockets==15.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
